no shows/TuBoleta/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.10
python-dateutil==2.8.2
